Fix using user specific path for `provider.wallet`
